+++ sdk/src/prelude.rs
@@ -8,7 +8,6 @@
 #![allow(ambiguous_glob_reexports)]
 #![doc(hidden)]
 
-pub use async_wsocket::ConnectionMode;
 pub use futures::StreamExt;
 pub use nostr::prelude::*;
 pub use nostr_database::prelude::*;
@@ -17,5 +16,7 @@ pub use nostr_gossip::prelude::*;
 pub use crate::client::{self, *};
 pub use crate::monitor::{self, *};
 pub use crate::policy::*;
+#[cfg(not(target_arch = "wasm32"))]
+pub use crate::proxy::{self, *};
 pub use crate::relay::{self, *};
 pub use crate::*;